refactor: 스토리북 폴더 구조 수정
