+++ src/stories/components/header/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Button } from './Button';
 import './header.css';
+import { Button } from '../button/Button';
 
 type User = {
   name: string;
