+++ src/stories/components/page/Page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Header } from './Header';
 import './page.css';
+import { Header } from '../header/Header';
 
 type User = {
   name: string;
